// lib/api/generated/types.gen.ts
export type UserResponse = {
    /**
     * Id
     *
     * Unique database ID of the user
     */
    id: number;
    /**
     * Email
     *
     * User's email address
     */
    email: string;
    /**
     * Tier
     *
     * Subscription tier level ('free', 'premium', etc.)
     */
    tier: string;
};